requirements.txt
requests>=2.31.0
google-generativeai>=0.4.0
rich>=13.7.0
asyncio>=3.4.3
argparse>=1.4.0
python-dotenv>=1.0.0
